visioncv/requirements.txt
fastmcp
numpy==2.1.1
Pillow==10.4.0
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
colorlog
